add SP for all real types for DP cases

+++ Codes/SimKern_FD3Dspher/FD3Dspher.090305.s2fr.PML/code/srcF/seis3d_metric.F90
@@ -28,7 +28,7 @@ implicit none
 
 !-----------------------------------------------------------------------------
 
-real,allocatable :: gx(:),gy(:),gz(:),gxsin(:),gxcot(:), &
+real(SP),allocatable :: gx(:),gy(:),gz(:),gxsin(:),gxcot(:), &
      x_xi(:),y_eta(:),z_zeta(:)
 character (len=SEIS_STRLEN) :: filenm
